// src/types/matchesPredictions.ts
// Generated by https://quicktype.io

import { Match } from "./matches";

export interface MatchPrediction {
  id: number;
  isDeleted: boolean;
  isTeamAwins: boolean;
  isTeamBwins: boolean;
  isAdraft: boolean;
  deletedAt: null;
  match: Match;
  transaction: Transaction;
}

export interface Transaction {
  id: number;
  ammount: number;
  type: string;
  creditCardCode: string;
  isDeleted: boolean;
}